implement beql and fix likely ops
I wasn't aware that likely branches nulify the instruction at the delay
slot if the branch if not taken!

I had to troubleshoot a ton of assembly to realize that, but once I saw
that an iterator variable was incremented twice I wondered, what the
heck, this doesn't make any sense, and then I read in the documentation
about the NullifyCurrentInstruction() stuff.